requirements.txt
--index-url https://download.pytorch.org/whl/nightly/cpu
torch==2.6.0.dev20241219

openai-whisper~=20240930
psutil~=6.1.1